QUDA now compiles cleanly with gcc 4.9 - fixed compilationo error with more strict handling of string / stringstream interaction and all compiler warnings silenced

+++ lib/dslash_improved_staggered.cu
@@ -105,9 +105,9 @@ namespace quda {
     {
       const unsigned int max_shared = deviceProp.sharedMemPerBlock;
       // first try to advance block.y (number of right-hand sides per block)
-      if (param.block.y < nSrc && param.block.y < deviceProp.maxThreadsDim[1] &&
+      if (param.block.y < nSrc && param.block.y < (unsigned int)deviceProp.maxThreadsDim[1] &&
 	  sharedBytesPerThread()*param.block.x*param.block.y < max_shared &&
-	  (param.block.x*(param.block.y+1)) <= deviceProp.maxThreadsPerBlock) {
+	  (param.block.x*(param.block.y+1u)) <= (unsigned int)deviceProp.maxThreadsPerBlock) {
 	param.block.y++;
 	param.grid.y = (nSrc + param.block.y - 1) / param.block.y;
 	return true;
